test(fixtures): split typecheck registry assertions into their own file
